【RE】[AtCoder Beginner Contest 145]E - All-you-can-eat

+++ abc145/abc145_e.cpp
@@ -29,32 +29,32 @@ bool contain(set<char> &s, int a) { return s.find(a) != s.end(); }
 typedef priority_queue<ll, vector<ll>, greater<ll>> PQ_ASK;
 const int mod = 1000000007;
 
-vector<vector<int>> dynamic_programing(int n, int t, vector<P> menu) {
-    vector<vector<int>> dp(n + 1, vector<int>(t + 1, -1));
+vector<vector<ll>> dynamic_programing(ll n, ll t, vector<P> menu) {
+    vector<vector<ll>> dp(n + 1, vector<ll>(t + 1, -1));
     dp[0][0] = 0;
 
-    auto get = [&](int i, int j) {
+    auto get = [&](ll i, ll j) {
         return dp[i][j];
     };
 
-    auto set = [&](int i, int j, int v) {
+    auto set = [&](ll i, ll j, ll v) {
         if (i > n + 1) return;
         if (j > t + 1)return;
-        int value = get(i, j);
+        ll value = get(i, j);
         dp[i][j] = max(value, v);
     };
 
-    for (int i = 0; i < n; i++) {
+    for (ll i = 0; i < n; i++) {
         P p = menu[i];
-        for (int j = 0; j <= t; j++) {
-            int prev = get(i, j);
+        for (ll j = 0; j <= t; j++) {
+            ll prev = get(i, j);
             if (prev == -1) continue;
             set(i + 1, j + p.first, prev + p.second);
             set(i + 1, j, prev);
         }
     }
-    for (int i = 0; i <= n; i++) {
-        for (int j = 1; j <= t; j++) {
+    for (ll i = 0; i <= n; i++) {
+        for (ll j = 1; j <= t; j++) {
             if (dp[i][j] == -1) {
                 dp[i][j] = dp[i][j - 1];
             }
@@ -65,30 +65,30 @@ vector<vector<int>> dynamic_programing(int n, int t, vector<P> menu) {
 }
 
 int main() {
-    int n, t;
+    ll n, t;
     cin >> n >> t;
 
     vector<P> menu(n);
     rep(i, n) cin >> menu[i].first >> menu[i].second;
 
-    vector<vector<int>> dp1 = dynamic_programing(n, t - 1, menu);
+    vector<vector<ll>> dp1 = dynamic_programing(n, t - 1, menu);
     reverse(menu.begin(), menu.end());
-    vector<vector<int>> dp2 = dynamic_programing(n, t - 1, menu);
+    vector<vector<ll>> dp2 = dynamic_programing(n, t - 1, menu);
     reverse(menu.begin(), menu.end());
 
-    int ans = 0;
-    for (int i = 0; i < n; i++) {
+    ll ans = 0;
+    for (ll i = 0; i < n; i++) {
         assert(i < dp1.size());
         assert(i < dp2.size());
         // i を最後にとっておいた場合の
-        for (int bj = 0; bj <= t - 1; bj++) {
-            int aj = t - bj - 1;
+        for (ll bj = 0; bj <= t - 1; bj++) {
+            ll aj = t - bj - 1;
             assert(bj < dp1[i].size());
             assert(aj < dp2[i].size());
-            int bh = dp1[i][bj];
-            int ah = dp2[n - i - 1][aj];
-            int lh = menu[i].second;
-            int now = bh + ah + lh;
+            ll bh = dp1[i][bj];
+            ll ah = dp2[n - i - 1][aj];
+            ll lh = menu[i].second;
+            ll now = bh + ah + lh;
             cmax(ans, now);
         }
     }
